+++ src/test/scala/mesosphere/marathon/core/task/TaskTest.scala
@@ -7,6 +7,7 @@ import org.apache.mesos.{ Protos => MesosProtos }
 import org.scalatest.{ FunSuite, GivenWhenThen, Matchers }
 
 import scala.collection.immutable.Seq
+import org.scalatest.OptionValues._
 
 class TaskTest extends FunSuite with Mockito with GivenWhenThen with Matchers {
   import scala.collection.JavaConverters._
@@ -44,87 +45,81 @@ class TaskTest extends FunSuite with Mockito with GivenWhenThen with Matchers {
       MarathonTestHelper
         .runningTaskForApp(appWithoutIpAddress.id)
         .withAgentInfo(_.copy(host = host))
-        .withNetworking(Task.NetworkInfoList(networkWithOneIp1))
+        .withNetworkInfos(Seq(networkWithOneIp1))
 
     val taskWithMultipleNetworksAndOneIp =
       MarathonTestHelper
         .runningTaskForApp(appWithoutIpAddress.id)
         .withAgentInfo(_.copy(host = host))
-        .withNetworking(Task.NetworkInfoList(networkWithoutIp, networkWithOneIp1))
+        .withNetworkInfos(Seq(networkWithoutIp, networkWithOneIp1))
 
     val taskWithMultipleNetworkAndNoIp =
       MarathonTestHelper
         .runningTaskForApp(appWithoutIpAddress.id)
         .withAgentInfo(_.copy(host = host))
-        .withNetworking(Task.NetworkInfoList(networkWithoutIp, networkWithoutIp))
+        .withNetworkInfos(Seq(networkWithoutIp, networkWithoutIp))
 
     val taskWithOneNetworkAndMultipleIPs =
       MarathonTestHelper
         .runningTaskForApp(appWithoutIpAddress.id)
         .withAgentInfo(_.copy(host = host))
-        .withNetworking(Task.NetworkInfoList(networkWithMultipleIps))
+        .withNetworkInfos(Seq(networkWithMultipleIps))
 
     val taskWithMultipleNetworkAndMultipleIPs =
       MarathonTestHelper
         .runningTaskForApp(appWithoutIpAddress.id)
         .withAgentInfo(_.copy(host = host))
-        .withNetworking(Task.NetworkInfoList(networkWithOneIp1, networkWithOneIp2))
-  }
-
-  test("effectiveIpAddress returns the agent address for MarathonTask instances without their own IP addresses") {
-    val f = new Fixture
-    f.taskWithoutIp.effectiveIpAddress(f.appWithIpAddress) should equal (f.host)
-    f.taskWithoutIp.effectiveIpAddress(f.appWithoutIpAddress) should equal (f.host)
+        .withNetworkInfos(Seq(networkWithOneIp1, networkWithOneIp2))
   }
 
   test("effectiveIpAddress returns the container ip for MarathonTask instances with one NetworkInfo (if the app requests an IP)") {
     val f = new Fixture
-    f.taskWithOneIp.effectiveIpAddress(f.appWithIpAddress) should equal(f.ipString1)
+    f.taskWithOneIp.effectiveIpAddress(f.appWithIpAddress).value should equal(f.ipString1)
   }
 
   test("effectiveIpAddress returns the first container ip for for MarathonTask instances with multiple NetworkInfos (if the app requests an IP)") {
     val f = new Fixture
-    f.taskWithMultipleNetworksAndOneIp.effectiveIpAddress(f.appWithIpAddress) should equal (f.ipString1)
+    f.taskWithMultipleNetworksAndOneIp.effectiveIpAddress(f.appWithIpAddress).value should equal (f.ipString1)
   }
 
-  test("effectiveIpAddress falls back to the agent IP") {
+  test("effectiveIpAddress returns None if there is no ip") {
     val f = new Fixture
-    f.taskWithMultipleNetworkAndNoIp.effectiveIpAddress(f.appWithIpAddress) should equal (f.host)
+    f.taskWithMultipleNetworkAndNoIp.effectiveIpAddress(f.appWithIpAddress) should be (None)
   }
 
   test("effectiveIpAddress returns the agent ip for MarathonTask instances with one NetworkInfo (if the app does NOT request an IP)") {
     val f = new Fixture
-    f.taskWithOneIp.effectiveIpAddress(f.appWithoutIpAddress) should equal(f.host)
+    f.taskWithOneIp.effectiveIpAddress(f.appWithoutIpAddress).value should equal(f.host)
   }
 
-  test("ipAddresses returns an empty list for MarathonTask instances with no IPs") {
+  test("ipAddresses returns None for MarathonTask instances with no IPs") {
     val f = new Fixture
-    f.taskWithoutIp.ipAddresses should be (empty)
+    f.taskWithoutIp.launched.value.ipAddresses should be (None)
   }
 
   test("ipAddresses returns an empty list for MarathonTask instances with no IPs and multiple NetworkInfos") {
     val f = new Fixture
-    f.taskWithMultipleNetworkAndNoIp.ipAddresses should be (empty)
+    f.taskWithMultipleNetworkAndNoIp.launched.value.ipAddresses.value should be (empty)
   }
 
   test("ipAddresses returns all IPs for MarathonTask instances with multiple IPs") {
     val f = new Fixture
-    f.taskWithMultipleNetworkAndMultipleIPs.ipAddresses should equal(Seq(f.ipAddress1, f.ipAddress2))
+    f.taskWithMultipleNetworkAndMultipleIPs.launched.value.ipAddresses.value should equal(Seq(f.ipAddress1, f.ipAddress2))
   }
 
   test("ipAddresses returns all IPs for MarathonTask instances with multiple IPs and multiple NetworkInfos") {
     val f = new Fixture
-    f.taskWithMultipleNetworkAndMultipleIPs.ipAddresses should equal(Seq(f.ipAddress1, f.ipAddress2))
+    f.taskWithMultipleNetworkAndMultipleIPs.launched.value.ipAddresses.value should equal(Seq(f.ipAddress1, f.ipAddress2))
   }
 
   test("ipAddresses returns one IP for MarathonTask instances with one IP and one NetworkInfo") {
     val f = new Fixture
-    f.taskWithOneIp.ipAddresses should equal(Seq(f.ipAddress1))
+    f.taskWithOneIp.launched.value.ipAddresses.value should equal(Seq(f.ipAddress1))
   }
 
   test("ipAddresses returns one IP for MarathonTask instances with one IP and multiple NetworkInfo") {
     val f = new Fixture
-    f.taskWithMultipleNetworksAndOneIp.ipAddresses should equal(Seq(f.ipAddress1))
+    f.taskWithMultipleNetworksAndOneIp.launched.value.ipAddresses.value should equal(Seq(f.ipAddress1))
   }
 
 }